full-stack-forum-file-service/requirements.txt
Flask
Werkzeug
boto3
python-dotenv
botocore
